Speedind up topology detection when multiple tasks per node are used
Context:
Loading the topology is slow and appears to be a sequential operation
for tasks on the same node. To optimize this, only one rank per node
now discovers the topology and shares it with the other ranks on that node.

Before this patch, approximately 370 seconds were spent in hwloc_topology_load
when using 192 MPI tasks per node. With this patch, the time has been reduced
to around 6 seconds.

Signed-off-by: Jean-Yves Vet <vet@hpe.com>