Addressing a crash issue related to MPICH_OFI_NIC_POLICY=GPU
Context:
HPCAT is not compiled with the necessary libraries to support
MPICH_OFI_NIC_POLICY=GPU, as this would compromise the modularity of the tool.
As a consequence MPICH crashes when this feature is enabled.

This patch enables the tool to emulate NIC affinity in accordance with GPU NUMA
affinity when using Slingshot with Cray MPICH and the MPICH_OFI_NIC_POLICY
environment variable is set to GPU.

Signed-off-by: Jean-Yves Vet <vet@hpe.com>