Detecting binding and affinity issues, display hints in output
This patch enables HPCAT to automatically detect potential binding and
affinity issues that may lead to performance degradation. When such issues
are found, hints are displayed in the footer of the tabular output to guide
the user.

A new option, --disable-hints, can be used to turn off hint detection and display.

Signed-off-by: Jean-Yves Vet <vet@hpe.com>